+++ Sources/MapboxNavigation/CarPlayMapViewController.swift
@@ -291,7 +291,7 @@ open class CarPlayMapViewController: UIViewController {
             cameraOptions.pitch = 0
             navigationMapView.mapView.mapboxMap.setCamera(to: cameraOptions)
             
-            navigationMapView.fitCamera(to: activeRoute)
+            navigationMapView.fitCamera(to: [activeRoute])
         }
     }
 }